refactor: extract PRCard and FileTree components
